Fix: resolved NullPointerException during world regeneration

+++ src/main/java/de/buddelbubi/commands/subcommand/RegenerateCommand.java
@@ -16,18 +16,18 @@ public class RegenerateCommand extends SubCommand {
 
     public RegenerateCommand() {
         super("regenerate");
-        this.setAliases(new String[] {
-            "regenerate",
-            "reg",
-            "reset",
-            "regen"
+        this.setAliases(new String[]{
+                "regenerate",
+                "reg",
+                "reset",
+                "regen"
         });
     }
 
     @Override
     public CommandParameter[] getParameters() {
 
-        LinkedList < CommandParameter > parameters = new LinkedList < > ();
+        LinkedList<CommandParameter> parameters = new LinkedList<>();
         parameters.add(CommandParameter.newEnum(this.getName(), this.getAliases()));
         parameters.add(CommandParameter.newType("world", true, CommandParamType.STRING));
         return parameters.toArray(new CommandParameter[parameters.size()]);
@@ -37,41 +37,51 @@ public class RegenerateCommand extends SubCommand {
     public boolean execute(CommandSender sender, String arg1, String[] args) {
 
         if (!sender.hasPermission("worldmanager.admin") && !sender.hasPermission("worldmanager.regenerate")) {
-
             sender.sendMessage(WorldManager.prefix + "§cYou are lacking the permission §e'worldmanager.regenerate'.");
             return false;
-
         } else {
             if (args.length == 1 || args.length == 2) {
                 Level l = null;
                 if (args.length == 1) {
                     if (sender instanceof Player) {
                         l = ((Player) sender).getLevel();
-                    } else sender.sendMessage(WorldManager.prefix + "§cDo /worldmanager regenerate [World]*.");
-                } else
-                if (args.length == 2) {
-                    if (Server.getInstance().getLevelByName(args[1]) != null) {
-
-                        l = Server.getInstance().getLevelByName(args[1]);
-
                     } else {
+                        sender.sendMessage(WorldManager.prefix + "§cDo /worldmanager regenerate [World].");
+                        return false;
+                    }
+                } else if (args.length == 2) {
+                    l = Server.getInstance().getLevelByName(args[1]);
+                    if (l == null) {
                         sender.sendMessage(WorldManager.prefix + "§cThis world does not exist.");
                         return false;
                     }
                 }
+
                 long seed = l.getSeed();
                 Generator generator = l.getGenerator();
                 String name = l.getFolderName();
-                l.unload();
+
+                if (!Server.getInstance().unloadLevel(l)) {
+                    sender.sendMessage(WorldManager.prefix + "§cFailed to unload the world. Ensure no players are in the world.");
+                    return false;
+                }
+
                 File regionfolder = new File(Server.getInstance().getDataPath() + "worlds/" + name + "/region");
+                if (regionfolder.exists() && regionfolder.isDirectory()) {
+                    FileUtils.deleteDirectoryContents(regionfolder);
+                }
+
                 File worldfolder = new File(Server.getInstance().getDataPath() + "worlds/" + name);
-                FileUtils.deleteDirectoryContents(regionfolder);
-                FileUtils.deleteDirectoryContents(worldfolder);
-                worldfolder.delete();
+                if (worldfolder.exists() && worldfolder.isDirectory()) {
+                    FileUtils.deleteDirectoryContents(worldfolder);
+                    worldfolder.delete();
+                }
+
                 Server.getInstance().generateLevel(name, seed, generator.getClass());
                 sender.sendMessage(WorldManager.prefix + "§7World §8" + name + " §7regenerated.");
-
-            } else sender.sendMessage(WorldManager.prefix + "§cDo /worldmanager regenerate [World].");
+            } else {
+                sender.sendMessage(WorldManager.prefix + "§cDo /worldmanager regenerate [World].");
+            }
         }
         return false;
     }
